Fixed virtual keyboards lacking macros.memory instruction and refactored them
