+++ src/api/hardware-virtual/DefyWired.js
@@ -1,0 +1,257 @@
+export const DefyWired = {
+  device: {
+    info: {
+      vendor: "Dygma",
+      product: "Defy",
+      keyboardType: "wired",
+      displayName: "Dygma Defy wired",
+      urls: [
+        {
+          name: "Homepage",
+          url: "https://www.dygma.com/defy/"
+        }
+      ]
+    },
+    usb: {
+      vendorId: 13807,
+      productId: 16
+    },
+    keyboard: {
+      rows: 5,
+      columns: 16
+    },
+    keyboardUnderglow: {
+      rows: 2,
+      columns: 89
+    },
+    RGBWMode: true,
+    components: {},
+    instructions: {
+      en: {
+        updateInstructions:
+          "To update the firmware, the keyboard needs a special reset. When the countdown starts, press and hold the Escape key. Soon after the countdown finished, the Neuron's light should start a blue pulsing pattern, and the flashing will proceed. At this point, you should release the Escape key."
+      }
+    }
+  },
+  virtual: {
+    version: {
+      data: "v1.0.9beta 7622bb53 c9d9b7b-dirty",
+      eraseable: false
+    },
+    "keymap.custom": {
+      data:
+        "41 30 31 32 33 34 23785 0 0 22709 35 36 37 38 39 48 2091 20 26 8 21 23 23786 0 0 22733 28 24 12 18 19 47 43 4 22 7 9 10 2096 0 0 17152 11 13 14 15 45 52 0 29 27 6 25 5 0 0 0 0 17 16 54 55 56 49 49210 50013 49467 51514 52286 847 848 226 17494 228 41 42 49977 52030 230 231 100 4638 4639 4640 4641 4642 65535 0 0 65535 4643 4644 4645 4646 4647 65535 6678 6177 74 82 77 4144 65535 0 0 65535 65535 75 594 78 2096 65535 65535 1055 80 81 79 4182 65535 0 0 65535 2084 592 593 591 51 65535 65535 65535 65535 65535 6169 53863 0 0 0 0 65535 4396 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 335 336 65535 65535 65535 65535 65535 17492 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 2087 84 85 86 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 95 96 97 87 65535 65535 65535 65535 82 65535 65535 65535 65535 65535 65535 65535 92 93 94 54 65535 65535 65535 80 81 79 65535 65535 65535 65535 65535 65535 89 90 91 88 65535 49162 17451 65535 65535 65535 65535 65535 17492 65535 65535 65535 65535 65535 65535 98 99 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 2078 1055 1032 1057 2080 65535 65535 65535 65535 2083 1071 1072 2087 48 65535 65535 2078 2079 1056 2081 2082 65535 65535 65535 65535 2084 2085 2086 46 2096 65535 65535 65535 53 2084 2101 65535 65535 65535 65535 65535 1059 1076 1073 2094 1080 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 53862 53861 53856 53860 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 53853 53854 53857 53858 53859 53855 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 41 30 31 32 33 34 23785 0 0 22709 35 36 37 38 39 48 2091 20 26 8 21 23 23786 0 0 22733 28 24 12 18 19 47 43 4 22 7 49434 10 2096 0 0 17152 11 49438 14 15 45 52 225 29 27 6 25 5 0 0 0 0 17 16 54 55 56 229 226 224 44 42 17451 40 76 227 101 41 42 17452 40 44 228 230 17492 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 2087 84 85 86 65535 65535 65535 74 82 77 75 65535 65535 65535 65535 65535 95 96 97 87 65535 65535 70 80 81 79 78 65535 65535 65535 65535 65535 92 93 94 54 65535 65535 22710 22709 23786 23785 22733 65535 65535 65535 65535 65535 89 90 91 88 65535 65535 65535 65535 65535 17450 65535 65535 65535 65535 65535 65535 17452 65535 65535 98 99 65535 58 59 60 61 62 65535 65535 65535 65535 63 64 65 66 67 65535 65535 65535 65535 47 100 2148 65535 65535 65535 65535 1054 1077 1076 1073 1071 1072 65535 2078 1055 1056 2081 2082 65535 65535 65535 65535 2095 2083 2096 2085 2086 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 17450 65535 65535 65535 65535 65535 65535 17451 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 41 30 31 32 33 34 53 0 0 46 35 36 37 38 39 45 43 20 26 8 21 23 0 0 0 48 28 24 12 18 19 47 57 4 22 7 9 10 2096 0 0 49 11 13 14 15 51 52 225 29 27 6 25 5 0 0 0 0 17 16 54 55 56 229 224 227 49979 51514 51806 44 226 17494 0 847 848 42 52026 49469 231 230",
+      eraseable: true
+    },
+    "keymap.default": {
+      data:
+        "41 30 31 32 33 34 35 0 0 51518 37 38 39 45 46 42 43 20 26 8 21 23 128 0 0 28 24 12 18 19 47 48 57 4 22 7 9 10 129 0 0 11 13 14 15 51 52 229 49 29 27 6 25 5 0 0 0 0 17 16 54 55 56 229 224 227 42 76 44 40 226 225 79 81 80 40 17152 74 82 77 41 58 59 60 61 62 63 0 0 64 65 66 67 68 69 42 43 0 82 0 0 0 0 0 0 86 36 37 38 84 0 40 57 80 81 79 0 0 0 0 0 87 33 34 35 85 0 49 225 49 0 0 0 0 0 0 0 0 99 30 31 32 82 229 224 227 226 44 76 40 42 44 79 228 76 17492 39 44 80 81",
+      eraseable: false
+    },
+    "keymap.onlyCustom": {
+      data: "1",
+      eraseable: true
+    },
+    "settings.defaultLayer": {
+      data: "0",
+      eraseable: true
+    },
+    "settings.valid?": {
+      data: "true",
+      eraseable: false
+    },
+    "settings.version": {
+      data: "1",
+      eraseable: false
+    },
+    "settings.crc": {
+      data: "18462/18462",
+      eraseable: false
+    },
+    "eeprom.contents": {
+      data: "",
+      eraseable: false
+    },
+    "eeprom.free": {
+      data: "",
+      eraseable: false
+    },
+    "led.at": {
+      data: "",
+      eraseable: false
+    },
+    "led.setAll": {
+      data: "",
+      eraseable: false
+    },
+    "led.mode": {
+      data: "0",
+      eraseable: true
+    },
+    "led.brightness": {
+      data: "153",
+      eraseable: true
+    },
+    "led.theme": {
+      data: "",
+      eraseable: false
+    },
+    palette: {
+      data:
+        "10 0 0 245 166 0 145 0 210 131 0 35 93 178 0 33 206 0 25 2 0 147 114 80 0 0 0 0 125 0 235 19 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0",
+      eraseable: true
+    },
+    "colormap.map": {
+      data:
+        "2 0 0 0 0 0 2 2 0 0 0 0 0 2 2 0 0 0 2 0 1 6 0 0 0 0 0 1 1 5 1 1 2 2 1 2 0 0 0 0 0 2 2 0 0 0 0 0 2 2 0 0 0 2 0 1 2 0 0 0 0 0 2 2 5 1 1 2 2 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 4 4 4 4 4 0 0 6 6 1 5 2 5 1 6 6 1 2 2 2 1 7 7 7 6 7 0 5 1 1 0 1 1 1 1 1 6 4 4 4 0 0 6 6 1 1 2 1 7 6 6 1 2 2 2 1 7 6 6 6 7 1 6 1 1 0 1 1 2 2 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 7 7 7 7 7 7 7 7 7 7 7 7 7 7 6 7 7 7 2 7 6 6 6 7 2 2 2 7 7 7 7 6 7 6 6 0 6 1 1 1 1 6 6 6 1 0 0 0 6 6 6 1 0 1 0 6 6 6 1 0 0 0 6 0 0 1 1 1 2 2 1 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 7 7 6 6 6 6 6 6 7 2 2 2 2 2 6 7 1 1 1 1 1 7 7 7 4 5 4 7 0 0 0 0 4 4 0 4 7 7 6 6 6 6 6 7 2 2 2 2 2 6 7 1 1 1 1 1 6 7 4 4 5 5 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 15 15 15 15 15 15 15 15 15 5 4 2 3 15 15 15 1 1 1 1 1 1 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 15 2 0 0 0 0 0 2 2 0 0 0 0 0 2 2 0 0 0 0 0 1 7 0 0 0 0 0 7 7 7 7 4 7 7 7 2 0 0 0 0 0 2 2 0 0 0 0 0 2 2 0 0 0 0 0 1 7 0 0 0 0 0 7 7 7 7 2 7 7 7 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 4 6 6 6 6 6 6 6 6 6 3 2 3 1 6 6 7 2 2 2 1 6 6 5 5 5 5 5 6 6 6 6 7 6 6 6 6 1 1 1 1 6 6 6 1 0 0 0 6 6 6 1 0 1 0 6 6 6 1 0 0 0 6 0 0 6 6 2 6 6 6 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 7 6 3 3 3 3 3 8 6 6 6 1 1 1 6 6 1 1 1 1 1 6 6 6 6 6 6 6 6 6 6 8 7 6 6 6 6 3 3 3 3 3 6 7 7 1 1 1 1 6 6 1 1 1 1 1 6 6 6 6 6 6 6 6 6 6 6 4 6 6 6 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 4 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 3 6 6 6 6 6 3 3 6 6 6 6 6 3 3 6 6 6 6 6 3 3 6 6 6 6 6 3 3 4 3 3 3 3 1 3 6 6 6 6 6 3 3 6 6 6 6 6 3 3 6 6 6 6 6 3 3 6 6 6 6 6 3 3 4 3 3 5 5 7 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 6 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15",
+      eraseable: true
+    },
+    "idleleds.time_limit": {
+      data: "300",
+      eraseable: true
+    },
+    "hardware.version": {
+      data: "Dygma Defy",
+      eraseable: false
+    },
+    "hardware.side_power": {
+      data: "1",
+      eraseable: false
+    },
+    "hardware.side_ver": {
+      data: "left: 5\r\nright: 5 ",
+      eraseable: false
+    },
+    "hardware.sled_ver": {
+      data: "left: 115\r\nright: 115 ",
+      eraseable: false
+    },
+    "hardware.sled_current": {
+      data: "left: 63\r\nright: 63 ",
+      eraseable: false
+    },
+    "hardware.layout": {
+      data: "ISO",
+      eraseable: false
+    },
+    "hardware.joint": {
+      data: "127",
+      eraseable: false
+    },
+    "hardware.keyscan": {
+      data: "50",
+      eraseable: false
+    },
+    "hardware.crc_errors": {
+      data: "left: 0 right: 0 ",
+      eraseable: false
+    },
+    "hardware.firmware": {
+      data: "c9d9b7b",
+      eraseable: false
+    },
+    "hardware.chip_id": {
+      data: "VirtualDefyWiredVirtualDefyWired",
+      eraseable: false
+    },
+    "qukeys.holdTimeout": {
+      data: "250",
+      eraseable: true
+    },
+    "qukeys.overlapThreshold": {
+      data: "80",
+      eraseable: true
+    },
+    "superkeys.map": {
+      data:
+        "65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535 65535",
+      eraseable: true
+    },
+    "superkeys.waitfor": {
+      data: "500",
+      eraseable: true
+    },
+    "superkeys.timeout": {
+      data: "250",
+      eraseable: true
+    },
+    "superkeys.repeat": {
+      data: "20",
+      eraseable: true
+    },
+    "superkeys.holdstart": {
+      data: "200",
+      eraseable: true
+    },
+    "superkeys.overlap": {
+      data: "20",
+      eraseable: true
+    },
+    "macros.map": {
+      data:
+        "6 225 8 5 7 225 8 8 8 22 8 23 8 54 8 40 8 40 6 225 8 15 7 225 8 24 8 12 8 22 0 6 227 8 80 7 227 8 84 8 11 8 30 8 40 6 227 8 79 7 227 0 6 227 8 80 7 227 8 84 8 11 8 31 8 40 6 227 8 79 7 227 0 6 227 8 80 7 227 8 84 8 5 8 24 8 15 8 15 8 8 8 23 8 8 8 7 8 44 8 15 8 12 8 22 8 23 8 40 0 6 227 8 80 7 227 8 84 8 5 8 4 8 17 8 17 8 8 8 21 8 44 8 28 8 8 8 15 8 15 8 18 8 26 8 40 0 6 227 8 80 7 227 8 84 8 11 8 32 8 40 6 227 8 79 7 227 0 6 227 8 80 7 227 8 84 8 11 8 33 8 40 6 227 8 79 7 227 0 6 227 8 80 7 227 8 84 8 6 8 11 8 8 8 6 8 14 8 15 8 12 8 22 8 23 8 40 0 6 227 8 80 7 227 8 84 8 5 8 4 8 17 8 17 8 8 8 21 8 44 8 10 8 21 8 8 8 8 8 17 8 40 6 227 8 79 7 227 0 6 227 8 80 7 227 8 84 8 5 8 4 8 17 8 17 8 8 8 21 8 44 8 21 8 8 8 7 8 40 6 227 8 79 7 227 0 6 227 8 80 7 227 8 84 8 5 8 4 8 17 8 17 8 8 8 21 8 44 8 5 8 15 8 24 8 8 8 40 6 227 8 79 7 227 0 8 84 8 10 8 12 8 19 8 11 8 28 2 0 200 8 40 0 0 8 10 8 21 8 8 8 8 8 17 8 40 0 8 84 8 5 8 4 8 17 8 17 8 8 8 21 8 44 8 21 8 8 8 7 8 40 0 0 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255 255",
+      eraseable: true
+    },
+    "macros.trigger": {
+      data: "false",
+      eraseable: false
+    },
+    "macros.memory": {
+      data: "2048",
+      eraseable: false
+    },
+    help: {
+      data:
+        "version\r\nkeymap.custom\r\nkeymap.default\r\nkeymap.onlyCustom\r\nsettings.defaultLayer\r\nsettings.valid?\r\nsettings.version\r\nsettings.crc\r\neeprom.contents\r\neeprom.free\r\nled.at\r\nled.setAll\r\nled.mode\r\nled.brightness\r\nled.theme\r\npalette\r\ncolormap.map\r\nidleleds.time_limit\r\nhardware.version\r\nhardware.side_power\r\nhardware.side_ver\r\nhardware.sled_ver\r\nhardware.sled_current\r\nhardware.layout\r\nhardware.joint\r\nhardware.keyscan\r\nhardware.crc_errors\r\nhardware.firmware\r\nhardware.chip_id\r\nqukeys.holdTimeout\r\nqukeys.overlapThreshold\r\nsuperkeys.map\r\nsuperkeys.waitfor\r\nsuperkeys.timeout\r\nsuperkeys.repeat\r\nsuperkeys.holdstart\r\nsuperkeys.overlap\r\nmacros.map\r\nmacros.trigger\r\nhelp\r\nmouse.speed\r\nmouse.speedDelay\r\nmouse.accelSpeed\r\nmouse.accelDelay\r\nmouse.wheelSpeed\r\nmouse.wheelDelay\r\nmouse.speedLimit\r\nlayer.activate\r\nlayer.deactivate\r\nlayer.isActive\r\nlayer.moveTo\r\nlayer.state",
+      eraseable: false
+    },
+    "mouse.speed": {
+      data: "20",
+      eraseable: true
+    },
+    "mouse.speedDelay": {
+      data: "2",
+      eraseable: true
+    },
+    "mouse.accelSpeed": {
+      data: "20",
+      eraseable: true
+    },
+    "mouse.accelDelay": {
+      data: "2",
+      eraseable: true
+    },
+    "mouse.wheelSpeed": {
+      data: "2",
+      eraseable: true
+    },
+    "mouse.wheelDelay": {
+      data: "100",
+      eraseable: true
+    },
+    "mouse.speedLimit": {
+      data: "127",
+      eraseable: true
+    },
+    "layer.activate": {
+      data: "false",
+      eraseable: false
+    },
+    "layer.deactivate": {
+      data: "false",
+      eraseable: false
+    },
+    "layer.isActive": {
+      data: "0",
+      eraseable: false
+    },
+    "layer.moveTo": {
+      data: "false",
+      eraseable: false
+    },
+    "layer.state": {
+      data: "1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0",
+      eraseable: false
+    }
+  }
+};
