+++ src/api/hardware-virtual/index.js
@@ -1,0 +1,8 @@
+import { RaiseANSI } from "./RaiseANSI";
+import { RaiseISO } from "./RaiseISO";
+import { DefyWired } from "./DefyWired";
+import { DefyWireless } from "./DefyWireless";
+
+const enumerator = [RaiseANSI, RaiseISO, DefyWireless];
+
+export { RaiseISO, RaiseANSI, DefyWired, DefyWireless, enumerator };
